+++ packages/cli/src/commands/login/index.ts
@@ -1,7 +1,10 @@
+import log from 'npmlog';
 import type { Command } from 'commander';
+import { password } from '@inquirer/prompts';
 
 import config from '@/config';
-import { login } from '@/request/api/login';
+import { login } from '@/request';
+import { output } from '@/utils';
 
 const COMMAND_NAME = 'login';
 
@@ -11,29 +14,34 @@ const COMMAND_NAME = 'login';
 export function registerCommand(program: Command) {
   return program
     .command(COMMAND_NAME)
-    .description('登录知晓云应用，登录后可获取知晓云相关的数据')
-    .argument('<clientSecret>', '知晓云应用的密钥')
-    .argument('[clientId]', '知晓云应用的 clint_id, 默认从 rc 文件中读取')
-    .action(async (clientId, clientSecret) => {
-      console.log('[command/login] login options --> ', clientId, clientSecret);
-      const id = clientId || config.client_id;
-      if (!id) {
-        program.error('未从本地配置中读取到 clint_id, 请配置或传入 clint_id!');
-      }
-
+    .description('Log in to MinCloud for data access')
+    .argument('[clientId]', 'Log in to MinCloud for data access')
+    .action(async clientId => {
       try {
+        const clientSecret = await password({
+          message: 'Please enter your app client secret:',
+        });
+
+        const id = clientId || config.client_id;
+        if (!id) {
+          log.error(
+            'login',
+            'Client ID not found! please configure or pass the client ID!',
+          );
+        }
+
         await login({
           clientId: id,
           clientSecret: clientSecret,
         });
-      } catch (error) {
-        // console.log('error', error);
 
-        if (error instanceof Error) {
-          program.error(`[command/login] 登录失败: ${error.message}`);
-        } else {
-          program.error('[command/login] 登录失败，请稍后再试');
-        }
+        output('Login successful!');
+      } catch (error) {
+        log.error(
+          'login',
+          'Login failed:',
+          error instanceof Error ? error.message : error,
+        );
       }
     });
 }