+++ packages/cli/src/request/utils.ts
@@ -1,12 +1,23 @@
 import type { AxiosError } from 'axios';
+import { isPlainObject } from 'lodash';
+
 import { ERROR_MESSAGE } from '../config/api';
 
-export function transformErrorMessage(error: AxiosError) {
-  const status = error.response?.status;
+export function transformErrorMessage(
+  error: AxiosError<{ status: string; error_msg: string }>,
+) {
+  const { status, data } = error.response || {};
 
   // 设置错误信息
   if (status) {
-    const message = error.response?.data || ERROR_MESSAGE[status];
+    let message = ERROR_MESSAGE[status];
+
+    if (typeof data === 'string') {
+      message = data;
+    } else if (isPlainObject(data) && data?.error_msg) {
+      message = data.error_msg;
+    }
+
     if (message) {
       error.message = message;
     }